requirements.txt
streamlit
python-docx
transformers
torch